pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = ""
authors = ["misdocumeno <misdocumeno@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-a2s = "^1.3.0"
python-dotenv = "^1.0.1"
py7zr = "^0.22.0"
requests = "^2.32.3"
vpk = "^1.4.0"
vdf = "^3.4"
configargparse = "^1.7"
fastapi = "^0.117.1"
slowapi = "^0.1.9"
uvicorn = "^0.37.0"
python-valve = "^0.2.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pyright = "^1.1.403"
poethepoet = "^0.36.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
typecheck = "pyright runner --pythonversion 3.10"
